Add user feedback for invalid path (#95)
* Add user feedback for invalid path

Now if the user provides an invalid path to monitor they'll receive an error message informing them of that.

* Remove printing stacktrace

+++ Gui/Controllers/HomeController.cs
@@ -402,7 +402,7 @@ namespace AttackSurfaceAnalyzer.Gui.Controllers
                 FilterLocation = "filters.json"
             };
             AttackSurfaceAnalyzerCLI.ClearMonitors();
-            return Json(AttackSurfaceAnalyzerCLI.RunGuiMonitorCommand(opts));
+            return Json((int)AttackSurfaceAnalyzerCLI.RunGuiMonitorCommand(opts));
         }
 
         public ActionResult StopMonitoring()
